adding Knowledge Bases task
