+++ lambdas/llmFunctions/kbfunction.py
@@ -1,67 +1,23 @@
-import os
-import json
 import boto3
-
-import traceback
+import json
 
 
 region = boto3.session.Session().region_name
-KB_ID = os.environ["KB_ID"]
-
 
 def lambda_handler(event, context):
-    boto3_version = boto3.__version__
-    print(f"Boto3 version: {boto3_version}")
-    
     print(f"Event is: {event}")
-    event_body = json.loads(event["body"])
-    prompt = event_body["query"]
-    model_id = event_body["model_id"]
-    
-    response = ''
+
+    response = 'This function is not yet available. Please proceed to Task 8, where you will implement it.'
     status_code = 200
     
-    try:
-        model_arn = 'arn:aws:bedrock:'+region+'::foundation-model/'+model_id
-        print(f"Model arn: {model_arn}")
-        
-        response = retrieveAndGenerate(prompt, model_arn)["output"]["text"]
-        return {
-            'statusCode': status_code,
-            'headers': {
-                'Access-Control-Allow-Origin': '*',
-                'Access-Control-Allow-Headers': 'Content-Type,X-Amz-Date,Authorization,X-Api-Key,X-Amz-Security-Token',
-                'Access-Control-Allow-Methods': 'OPTIONS,POST'
-            },
-            'body': json.dumps({'answer': response})
-        }
-            
-    except Exception as e:
-        print(f"An unexpected error occurred: {str(e)}")
-        stack_trace = traceback.format_exc()
-        print(stack_trace)
-        return {
-            'statusCode': status_code,
-            'headers': {
-                'Access-Control-Allow-Origin': '*',
-                'Access-Control-Allow-Headers': 'Content-Type,X-Amz-Date,Authorization,X-Api-Key,X-Amz-Security-Token',
-                'Access-Control-Allow-Methods': 'OPTIONS,POST'
-            },
-            'body': json.dumps({'error': str(e)})
-        }
 
-def retrieveAndGenerate(prompt, model_arn):
-    bedrock_agent_runtime = boto3.client(
-            service_name = "bedrock-agent-runtime")
-    return bedrock_agent_runtime.retrieve_and_generate(
-        input={
-            'text': prompt
+    return {
+        'statusCode': status_code,
+        'headers': {
+            'Access-Control-Allow-Origin': '*',
+            'Access-Control-Allow-Headers': 'Content-Type,X-Amz-Date,Authorization,X-Api-Key,X-Amz-Security-Token',
+            'Access-Control-Allow-Methods': 'OPTIONS,POST'
         },
-        retrieveAndGenerateConfiguration={
-            'type': 'KNOWLEDGE_BASE',
-            'knowledgeBaseConfiguration': {
-                'knowledgeBaseId': KB_ID,
-                'modelArn': model_arn
-                }
-            }
-    )
+        'body': json.dumps({'answer': response})
+    }
+
