+++ lambdas/llmFunctions/llmfunction.py
@@ -1,6 +1,5 @@
 import boto3
 import json
-import json
 import traceback
