+++ tools/solutions/kbfunction.py
@@ -2,51 +2,66 @@ import os
 import json
 import boto3
 
+import traceback
+
+
+region = boto3.session.Session().region_name
 KB_ID = os.environ["KB_ID"]
 
 
 def lambda_handler(event, context):
-    response = retrieveAndGenerate("What is latest federal interest rate?")["output"]["text"]
-    #response = retrieve("What is latest federal interest rate?", "JIXOVPTR6Z")["retrievalResults"]
-    print(f"Response:  {response}")  
-    return {
-        'statusCode': 200,
-        'headers': {
-            'Access-Control-Allow-Origin': '*',
-            'Access-Control-Allow-Headers': 'Content-Type,X-Amz-Date,Authorization,X-Api-Key,X-Amz-Security-Token',
-            'Access-Control-Allow-Methods': 'OPTIONS,POST'
-        },
-        'body': json.dumps({'answer': response})
-    }
+    boto3_version = boto3.__version__
+    print(f"Boto3 version: {boto3_version}")
+    
+    print(f"Event is: {event}")
+    event_body = json.loads(event["body"])
+    prompt = event_body["query"]
+    model_id = event_body["model_id"]
+    
+    response = ''
+    status_code = 200
+    
+    try:
+        model_arn = 'arn:aws:bedrock:'+region+'::foundation-model/'+model_id
+        print(f"Model arn: {model_arn}")
+        
+        response = retrieveAndGenerate(prompt, model_arn)["output"]["text"]
+        return {
+            'statusCode': status_code,
+            'headers': {
+                'Access-Control-Allow-Origin': '*',
+                'Access-Control-Allow-Headers': 'Content-Type,X-Amz-Date,Authorization,X-Api-Key,X-Amz-Security-Token',
+                'Access-Control-Allow-Methods': 'OPTIONS,POST'
+            },
+            'body': json.dumps({'answer': response})
+        }
+            
+    except Exception as e:
+        print(f"An unexpected error occurred: {str(e)}")
+        stack_trace = traceback.format_exc()
+        print(stack_trace)
+        return {
+            'statusCode': status_code,
+            'headers': {
+                'Access-Control-Allow-Origin': '*',
+                'Access-Control-Allow-Headers': 'Content-Type,X-Amz-Date,Authorization,X-Api-Key,X-Amz-Security-Token',
+                'Access-Control-Allow-Methods': 'OPTIONS,POST'
+            },
+            'body': json.dumps({'error': str(e)})
+        }
 
-
-def retrieveAndGenerate(input):
+def retrieveAndGenerate(prompt, model_arn):
     bedrock_agent_runtime = boto3.client(
             service_name = "bedrock-agent-runtime")
     return bedrock_agent_runtime.retrieve_and_generate(
         input={
-            'text': input
+            'text': prompt
         },
         retrieveAndGenerateConfiguration={
             'type': 'KNOWLEDGE_BASE',
             'knowledgeBaseConfiguration': {
                 'knowledgeBaseId': KB_ID,
-                'modelArn': 'arn:aws:bedrock:us-west-2::foundation-model/anthropic.claude-3-haiku-20240307-v1:0'
+                'modelArn': model_arn
                 }
             }
-        )
-        
-def retrieve(query, kbId, numberOfResults=5):
-    bedrock_agent_runtime = boto3.client(
-            service_name = "bedrock-agent-runtime")
-    return bedrock_agent_runtime.retrieve(
-        retrievalQuery= {
-            'text': query
-        },
-        knowledgeBaseId=kbId,
-        retrievalConfiguration= {
-            'vectorSearchConfiguration': {
-                'numberOfResults': numberOfResults
-            }
-        }
     )
